+++ controllers/sourcesControllers.js
@@ -1,0 +1,24 @@
+const Source = require("../models/sourceModel");
+const { catchAsyncErrors } = require("../middlewares/catchAsyncErrors");
+const logger = require("../logger");
+
+exports.createSource = catchAsyncErrors(async (req, res) => {
+    logger.info("You made a POST Request on Source creation Route");
+  
+    if (Array.isArray(req.body)) {
+      const sources = await Source.insertMany(req.body);
+      // await sources.save();
+      res
+        .status(201)
+        .send({ success: true, message: "Sources Created in Bulk successfully" });
+      logger.info(sources);
+    } else {
+      const source = new Source(req.body);
+      await source.save();
+      res
+        .status(201)
+        .send({ success: true, message: "Sources Created successfully" });
+      logger.info(source);
+    }
+  });
+
