+++ routes/Sources-routes.js
@@ -1,0 +1,8 @@
+const {createSource} = require('../controllers/sourcesControllers');                    
+const express = require('express');
+const router = express.Router();
+
+// Create a new source
+router.post('/',createSource);
+
+module.exports = router;
